tests/wpt/tests/css/css-overflow/parsing/continue-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Overflow: parsing continue with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-overflow-4/#propdef-continue">
<meta name="assert" content="continue supports the full grammar 'auto | discard | collapse', as well as '-webkit-legacy' for legacy compatibility.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("continue", 'auto');
test_valid_value("continue", 'discard');
test_valid_value("continue", 'collapse');
test_valid_value("continue", '-webkit-legacy');
</script>
</body>
</html>
